requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.4.0
fastapi==0.112.2
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.5
grpcio-status==1.62.3
h11==0.14.0
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
mutagen==1.47.0
packaging==24.1
pluggy==1.5.0
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.2
pytest==8.3.2
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
rich==13.7.1
rsa==4.9
sniffio==1.3.1
spotipy==2.24.0
SQLAlchemy==2.0.32
starlette==0.38.2
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
websockets==13.0
yarl==1.9.4
yt-dlp==2024.8.6
